Super User App Implementation

+++ app/controllers/dashboard_controller.rb
@@ -24,8 +24,8 @@ class DashboardController < ApplicationController
   def create_program_admin
     @current_program = Program.find(current_user.program_admin.program_id)
 
-    @admin_user = User.create(email: program_admin_params[:email],   first_name: program_admin_params[:first_name],
-    last_name:  program_admin_params[:last_name], password: "password")
+    @admin_user = User.create(email: program_admin_params[:email], first_name: program_admin_params[:first_name],
+                              last_name: program_admin_params[:last_name], password: "password")
     @program_admin = ProgramAdmin.new(user_id: @admin_user.id, program_id: @current_program.id, created_by_admin_id: current_user.id)
 
     if @program_admin.save
@@ -37,9 +37,13 @@ class DashboardController < ApplicationController
   end
 
   def delete_program_admin
+    return unless program_admin.super_user?
+
     ActiveRecord::Base.transaction do
       program_admin = ProgramAdmin.find(params[:id])
       user = program_admin.admin
+      return unless program_admin.super_user? == false
+
       program_admin.destroy!
       user.delete
     end
